+++ .prettierrc.mjs
@@ -1,2 +1,2 @@
 /** @type {import("prettier").Config} */
-export default "@nesvel/prettier-config";
+export default '@nesvel/prettier-config';