+++ __tests__/setup.d.ts
@@ -2,7 +2,7 @@
  * Type declarations for Vitest in @abdokouta/logger
  */
 
-declare module "vitest" {
+declare module 'vitest' {
   interface Assertion<T = any> {}
   interface AsymmetricMatchersContaining {}
 }