+++ src/types/index.ts
@@ -1,7 +1,7 @@
 // Managers
-export type { DriverCreator } from "./driver-creator.type";
+export type { DriverCreator } from './driver-creator.type';
 
 // Registry
-export type { BaseRegistryOptions } from "./base-registry-options.type";
-export type { Collection } from "./collection.type";
-export type { ValidationResult } from "./validation-result.type";
+export type { BaseRegistryOptions } from './base-registry-options.type';
+export type { Collection } from './collection.type';
+export type { ValidationResult } from './validation-result.type';